requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dbus-python==1.2.18
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.95.1
greenlet==1.1.3.post0
# gyp==0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
# PyGObject==3.42.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.4.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
influxdb_client
